requirements.txt
Flask
pandas
openpyxl
Pillow
gspread
oauth2client
gunicorn
werkzeug
